requirements.txt
awslambdaric
Pillow
